Create DataCampCourse OutlineWorkspace Ready The curious case of missing values
We need to first find out which variables have missing values, and then see what is the best way to handle these missing values. The way to handle a missing value can depend on the number of missing values, the type of variable and the expected importance of those variables.